data_rate: Move call_repeatedly to be used in serial_link
